feat(desktop): bridge MCP per-tab and surface per-server status
Extract `createMcpRuntime` from `chat.tsx` into a non-tsx module
`cli/commands/mcp-runtime.ts` and import from both `chat` and `desktop`.
chat.tsx re-exports the public types for backwards compat.

In `desktop.ts`:

- Each `Tab` now owns an optional `mcpRuntime` plus a `mcpStatuses` map.
  `bridgeTabMcp` lazily constructs the runtime, wires a lifecycle sink
  that updates the per-spec status, and calls `reloadFromConfig` against
  the tab's loop so MCP tools register into the live prefix.
- `closeTab` calls `runtime.closeAll()` so MCP children exit cleanly.
- `mcp_specs_add` / `mcp_specs_remove` trigger `bridgeTabMcp` so live
  config edits reach the runtime without restarting the tab.
- `$mcp_specs` now carries `status` + optional `statusReason` /
  `toolCount` per spec. `bridged` stays as a coarse "all connected"
  flag.

The right-panel Tools tab renders the per-spec dot — green for
connected, yellow pending for handshake / configured, grey for failed /
disabled — and the row text shows the live state (tool count, failure
reason).

Closes #753.
